@@ -1,15 +1,32 @@
-module github.com/mholt/archiver/v3
+module github.com/mholt/archiver/v4
 
-go 1.13
+go 1.22.2
+
+toolchain go1.23.2
 
 require (
-	github.com/andybalholm/brotli v1.0.4
-	github.com/dsnet/compress v0.0.2-0.20210315054119-f66993602bf5
-	github.com/golang/snappy v0.0.2
-	github.com/klauspost/compress v1.11.4
-	github.com/klauspost/pgzip v1.2.5
-	github.com/nwaples/rardecode v1.1.0
-	github.com/pierrec/lz4/v4 v4.1.2
-	github.com/ulikunitz/xz v0.5.9
-	github.com/xi2/xz v0.0.0-20171230120015-48954b6210f8
+	github.com/andybalholm/brotli v1.1.1
+	github.com/dsnet/compress v0.0.2-0.20230904184137-39efe44ab707
+	github.com/klauspost/compress v1.17.11
+	github.com/klauspost/pgzip v1.2.6
+	github.com/nwaples/rardecode/v2 v2.0.0-beta.4.0.20241112120701-034e449c6e78
+	github.com/therootcompany/xz v1.0.1
+	github.com/ulikunitz/xz v0.5.12
+)
+
+require (
+	github.com/STARRY-S/zip v0.2.1
+	github.com/bodgit/sevenzip v1.6.0
+	github.com/pierrec/lz4/v4 v4.1.21
+	github.com/sorairolake/lzip-go v0.3.5
+	golang.org/x/text v0.20.0
+)
+
+require (
+	github.com/bodgit/plumbing v1.3.0 // indirect
+	github.com/bodgit/windows v1.0.1 // indirect
+	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/hashicorp/go-multierror v1.1.1 // indirect
+	github.com/hashicorp/golang-lru/v2 v2.0.7 // indirect
+	go4.org v0.0.0-20230225012048-214862532bf5 // indirect
 )
