@@ -2,676 +2,352 @@ package archiver
 
 import (
 	"bytes"
-	"compress/flate"
+	"context"
+	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
+	"io/fs"
 	"log"
-	"os"
 	"path"
-	"path/filepath"
 	"strings"
+
+	szip "github.com/STARRY-S/zip"
 
 	"github.com/dsnet/compress/bzip2"
 	"github.com/klauspost/compress/zip"
 	"github.com/klauspost/compress/zstd"
 	"github.com/ulikunitz/xz"
+	"golang.org/x/text/encoding"
+	"golang.org/x/text/encoding/charmap"
+	"golang.org/x/text/encoding/japanese"
+	"golang.org/x/text/encoding/korean"
+	"golang.org/x/text/encoding/simplifiedchinese"
+	"golang.org/x/text/encoding/traditionalchinese"
+	"golang.org/x/text/encoding/unicode"
 )
 
-// ZipCompressionMethod Compression type
-type ZipCompressionMethod uint16
+func init() {
+	RegisterFormat(Zip{})
 
-// Compression methods.
-// see https://pkware.cachefly.net/webdocs/casestudies/APPNOTE.TXT.
-// Note LZMA: Disabled - because 7z isn't able to unpack ZIP+LZMA ZIP+LZMA2 archives made this way - and vice versa.
-const (
-	Store   ZipCompressionMethod = 0
-	Deflate ZipCompressionMethod = 8
-	BZIP2   ZipCompressionMethod = 12
-	LZMA    ZipCompressionMethod = 14
-	ZSTD    ZipCompressionMethod = 93
-	XZ      ZipCompressionMethod = 95
-)
-
-// Zip provides facilities for operating ZIP archives.
-// See https://pkware.cachefly.net/webdocs/casestudies/APPNOTE.TXT.
-type Zip struct {
-	// The compression level to use, as described
-	// in the compress/flate package.
-	CompressionLevel int
-
-	// Whether to overwrite existing files; if false,
-	// an error is returned if the file exists.
-	OverwriteExisting bool
-
-	// Whether to make all the directories necessary
-	// to create a zip archive in the desired path.
-	MkdirAll bool
-
-	// If enabled, selective compression will only
-	// compress files which are not already in a
-	// compressed format; this is decided based
-	// simply on file extension.
-	SelectiveCompression bool
-
-	// A single top-level folder can be implicitly
-	// created by the Archive or Unarchive methods
-	// if the files to be added to the archive
-	// or the files to be extracted from the archive
-	// do not all have a common root. This roughly
-	// mimics the behavior of archival tools integrated
-	// into OS file browsers which create a subfolder
-	// to avoid unexpectedly littering the destination
-	// folder with potentially many files, causing a
-	// problematic cleanup/organization situation.
-	// This feature is available for both creation
-	// and extraction of archives, but may be slightly
-	// inefficient with lots and lots of files,
-	// especially on extraction.
-	ImplicitTopLevelFolder bool
-
-	// Strip number of leading paths. This feature is available
-	// only during unpacking of the entire archive.
-	StripComponents int
-
-	// If true, errors encountered during reading
-	// or writing a single file will be logged and
-	// the operation will continue on remaining files.
-	ContinueOnError bool
-
-	// Compression algorithm
-	FileMethod ZipCompressionMethod
-	zw         *zip.Writer
-	zr         *zip.Reader
-	ridx       int
-	//decinitialized bool
-}
-
-// CheckExt ensures the file extension matches the format.
-func (*Zip) CheckExt(filename string) error {
-	if !strings.HasSuffix(filename, ".zip") {
-		return fmt.Errorf("filename must have a .zip extension")
-	}
-	return nil
-}
-
-// Registering a global decompressor is not reentrant and may panic
-func registerDecompressor(zr *zip.Reader) {
-	// register zstd decompressor
-	zr.RegisterDecompressor(uint16(ZSTD), func(r io.Reader) io.ReadCloser {
-		zr, err := zstd.NewReader(r)
-		if err != nil {
-			return nil
-		}
-		return zr.IOReadCloser()
+	// TODO: What about custom flate levels too
+	zip.RegisterCompressor(ZipMethodBzip2, func(out io.Writer) (io.WriteCloser, error) {
+		return bzip2.NewWriter(out, &bzip2.WriterConfig{ /*TODO: Level: z.CompressionLevel*/ })
 	})
-	zr.RegisterDecompressor(uint16(BZIP2), func(r io.Reader) io.ReadCloser {
+	zip.RegisterCompressor(ZipMethodZstd, func(out io.Writer) (io.WriteCloser, error) {
+		return zstd.NewWriter(out)
+	})
+	zip.RegisterCompressor(ZipMethodXz, func(out io.Writer) (io.WriteCloser, error) {
+		return xz.NewWriter(out)
+	})
+
+	zip.RegisterDecompressor(ZipMethodBzip2, func(r io.Reader) io.ReadCloser {
 		bz2r, err := bzip2.NewReader(r, nil)
 		if err != nil {
 			return nil
 		}
 		return bz2r
 	})
-	zr.RegisterDecompressor(uint16(XZ), func(r io.Reader) io.ReadCloser {
+	zip.RegisterDecompressor(ZipMethodZstd, func(r io.Reader) io.ReadCloser {
+		zr, err := zstd.NewReader(r)
+		if err != nil {
+			return nil
+		}
+		return zr.IOReadCloser()
+	})
+	zip.RegisterDecompressor(ZipMethodXz, func(r io.Reader) io.ReadCloser {
 		xr, err := xz.NewReader(r)
 		if err != nil {
 			return nil
 		}
-		return ioutil.NopCloser(xr)
+		return io.NopCloser(xr)
 	})
 }
 
-// CheckPath ensures the file extension matches the format.
-func (*Zip) CheckPath(to, filename string) error {
-	to, _ = filepath.Abs(to) //explicit the destination folder to prevent that 'string.HasPrefix' check can be 'bypassed' when no destination folder is supplied in input
-	dest := filepath.Join(to, filename)
-	//prevent path traversal attacks
-	if !strings.HasPrefix(dest, to) {
-		return &IllegalPathError{AbsolutePath: dest, Filename: filename}
-	}
-	return nil
+type Zip struct {
+	// Only compress files which are not already in a
+	// compressed format (determined simply by examining
+	// file extension).
+	SelectiveCompression bool
+
+	// The method or algorithm for compressing stored files.
+	Compression uint16
+
+	// If true, errors encountered during reading or writing
+	// a file within an archive will be logged and the
+	// operation will continue on remaining files.
+	ContinueOnError bool
+
+	// For files in zip archives that do not have UTF-8
+	// encoded filenames and comments, specify the character
+	// encoding here.
+	TextEncoding string
 }
 
-// Archive creates a .zip file at destination containing
-// the files listed in sources. The destination must end
-// with ".zip". File paths can be those of regular files
-// or directories. Regular files are stored at the 'root'
-// of the archive, and directories are recursively added.
-func (z *Zip) Archive(sources []string, destination string) error {
-	err := z.CheckExt(destination)
+func (z Zip) Extension() string { return ".zip" }
+
+func (z Zip) Match(_ context.Context, filename string, stream io.Reader) (MatchResult, error) {
+	var mr MatchResult
+
+	// match filename
+	if strings.Contains(strings.ToLower(filename), z.Extension()) {
+		mr.ByName = true
+	}
+
+	// match file header
+	buf, err := readAtMost(stream, len(zipHeader))
 	if err != nil {
-		return fmt.Errorf("checking extension: %v", err)
+		return mr, err
 	}
-	if !z.OverwriteExisting && fileExists(destination) {
-		return fmt.Errorf("file already exists: %s", destination)
-	}
+	mr.ByStream = bytes.Equal(buf, zipHeader)
 
-	// make the folder to contain the resulting archive
-	// if it does not already exist
-	destDir := filepath.Dir(destination)
-	if z.MkdirAll && !fileExists(destDir) {
-		err := mkdir(destDir, 0755)
-		if err != nil {
-			return fmt.Errorf("making folder for destination: %v", err)
-		}
-	}
-
-	out, err := os.Create(destination)
-	if err != nil {
-		return fmt.Errorf("creating %s: %v", destination, err)
-	}
-	defer out.Close()
-
-	err = z.Create(out)
-	if err != nil {
-		return fmt.Errorf("creating zip: %v", err)
-	}
-	defer z.Close()
-
-	var topLevelFolder string
-	if z.ImplicitTopLevelFolder && multipleTopLevels(sources) {
-		topLevelFolder = folderNameFromFileName(destination)
-	}
-
-	for _, source := range sources {
-		err := z.writeWalk(source, topLevelFolder, destination)
-		if err != nil {
-			return fmt.Errorf("walking %s: %v", source, err)
-		}
-	}
-
-	return nil
+	return mr, nil
 }
 
-// Unarchive unpacks the .zip file at source to destination.
-// Destination will be treated as a folder name.
-func (z *Zip) Unarchive(source, destination string) error {
-	if !fileExists(destination) && z.MkdirAll {
-		err := mkdir(destination, 0755)
-		if err != nil {
-			return fmt.Errorf("preparing destination: %v", err)
-		}
-	}
+func (z Zip) Archive(ctx context.Context, output io.Writer, files []FileInfo) error {
+	zw := zip.NewWriter(output)
+	defer zw.Close()
 
-	file, err := os.Open(source)
-	if err != nil {
-		return fmt.Errorf("opening source file: %v", err)
-	}
-	defer file.Close()
-
-	fileInfo, err := file.Stat()
-	if err != nil {
-		return fmt.Errorf("statting source file: %v", err)
-	}
-
-	err = z.Open(file, fileInfo.Size())
-	if err != nil {
-		return fmt.Errorf("opening zip archive for reading: %v", err)
-	}
-	defer z.Close()
-
-	// if the files in the archive do not all share a common
-	// root, then make sure we extract to a single subfolder
-	// rather than potentially littering the destination...
-	if z.ImplicitTopLevelFolder {
-		files := make([]string, len(z.zr.File))
-		for i := range z.zr.File {
-			files[i] = z.zr.File[i].Name
-		}
-		if multipleTopLevels(files) {
-			destination = filepath.Join(destination, folderNameFromFileName(source))
-		}
-	}
-
-	for {
-		err := z.extractNext(destination)
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			if z.ContinueOnError || IsIllegalPathError(err) {
-				log.Printf("[ERROR] Reading file in zip archive: %v", err)
-				continue
-			}
-			return fmt.Errorf("reading file in zip archive: %v", err)
-		}
-	}
-
-	return nil
-}
-
-func (z *Zip) extractNext(to string) error {
-	f, err := z.Read()
-	if err != nil {
-		return err // don't wrap error; calling loop must break on io.EOF
-	}
-	defer f.Close()
-
-	header, ok := f.Header.(zip.FileHeader)
-	if !ok {
-		return fmt.Errorf("expected header to be zip.FileHeader but was %T", f.Header)
-	}
-
-	errPath := z.CheckPath(to, header.Name)
-	if errPath != nil {
-		return fmt.Errorf("checking path traversal attempt: %v", errPath)
-	}
-
-	if z.StripComponents > 0 {
-		if strings.Count(header.Name, "/") < z.StripComponents {
-			return nil // skip path with fewer components
-		}
-
-		for i := 0; i < z.StripComponents; i++ {
-			slash := strings.Index(header.Name, "/")
-			header.Name = header.Name[slash+1:]
-		}
-	}
-	return z.extractFile(f, to, &header)
-}
-
-func (z *Zip) extractFile(f File, to string, header *zip.FileHeader) error {
-	to = filepath.Join(to, header.Name)
-
-	// if a directory, no content; simply make the directory and return
-	if f.IsDir() {
-		return mkdir(to, f.Mode())
-	}
-
-	// do not overwrite existing files, if configured
-	if !z.OverwriteExisting && fileExists(to) {
-		return fmt.Errorf("file already exists: %s", to)
-	}
-
-	// extract symbolic links as symbolic links
-	if isSymlink(header.FileInfo()) {
-		// symlink target is the contents of the file
-		buf := new(bytes.Buffer)
-		_, err := io.Copy(buf, f)
-		if err != nil {
-			return fmt.Errorf("%s: reading symlink target: %v", header.Name, err)
-		}
-		return writeNewSymbolicLink(to, strings.TrimSpace(buf.String()))
-	}
-
-	return writeNewFile(to, f, f.Mode())
-}
-
-func (z *Zip) writeWalk(source, topLevelFolder, destination string) error {
-	sourceInfo, err := os.Stat(source)
-	if err != nil {
-		return fmt.Errorf("%s: stat: %v", source, err)
-	}
-	destAbs, err := filepath.Abs(destination)
-	if err != nil {
-		return fmt.Errorf("%s: getting absolute path of destination %s: %v", source, destination, err)
-	}
-
-	return filepath.Walk(source, func(fpath string, info os.FileInfo, err error) error {
-		handleErr := func(err error) error {
-			if z.ContinueOnError {
-				log.Printf("[ERROR] Walking %s: %v", fpath, err)
-				return nil
-			}
+	for i, file := range files {
+		if err := z.archiveOneFile(ctx, zw, i, file); err != nil {
 			return err
 		}
-		if err != nil {
-			return handleErr(fmt.Errorf("traversing %s: %v", fpath, err))
+	}
+
+	return nil
+}
+
+func (z Zip) ArchiveAsync(ctx context.Context, output io.Writer, jobs <-chan ArchiveAsyncJob) error {
+	zw := zip.NewWriter(output)
+	defer zw.Close()
+
+	var i int
+	for job := range jobs {
+		job.Result <- z.archiveOneFile(ctx, zw, i, job.File)
+		i++
+	}
+
+	return nil
+}
+
+func (z Zip) archiveOneFile(ctx context.Context, zw *zip.Writer, idx int, file FileInfo) error {
+	if err := ctx.Err(); err != nil {
+		return err // honor context cancellation
+	}
+
+	hdr, err := zip.FileInfoHeader(file)
+	if err != nil {
+		return fmt.Errorf("getting info for file %d: %s: %w", idx, file.Name(), err)
+	}
+	hdr.Name = file.NameInArchive // complete path, since FileInfoHeader() only has base name
+	if hdr.Name == "" {
+		hdr.Name = file.Name() // assume base name of file I guess
+	}
+
+	// customize header based on file properties
+	if file.IsDir() {
+		if !strings.HasSuffix(hdr.Name, "/") {
+			hdr.Name += "/" // required
 		}
-		if info == nil {
-			return handleErr(fmt.Errorf("%s: no file info", fpath))
+		hdr.Method = zip.Store
+	} else if z.SelectiveCompression {
+		// only enable compression on compressable files
+		ext := strings.ToLower(path.Ext(hdr.Name))
+		if _, ok := compressedFormats[ext]; ok {
+			hdr.Method = zip.Store
+		} else {
+			hdr.Method = z.Compression
+		}
+	} else {
+		hdr.Method = z.Compression
+	}
+
+	w, err := zw.CreateHeader(hdr)
+	if err != nil {
+		return fmt.Errorf("creating header for file %d: %s: %w", idx, file.Name(), err)
+	}
+
+	// directories have no file body
+	if file.IsDir() {
+		return nil
+	}
+	if err := openAndCopyFile(file, w); err != nil {
+		return fmt.Errorf("writing file %d: %s: %w", idx, file.Name(), err)
+	}
+
+	return nil
+}
+
+// Extract extracts files from z, implementing the Extractor interface. Uniquely, however,
+// sourceArchive must be an io.ReaderAt and io.Seeker, which are oddly disjoint interfaces
+// from io.Reader which is what the method signature requires. We chose this signature for
+// the interface because we figure you can Read() from anything you can ReadAt() or Seek()
+// with. Due to the nature of the zip archive format, if sourceArchive is not an io.Seeker
+// and io.ReaderAt, an error is returned.
+func (z Zip) Extract(ctx context.Context, sourceArchive io.Reader, handleFile FileHandler) error {
+	sra, ok := sourceArchive.(seekReaderAt)
+	if !ok {
+		return fmt.Errorf("input type must be an io.ReaderAt and io.Seeker because of zip format constraints")
+	}
+
+	size, err := streamSizeBySeeking(sra)
+	if err != nil {
+		return fmt.Errorf("determining stream size: %w", err)
+	}
+
+	zr, err := zip.NewReader(sra, size)
+	if err != nil {
+		return err
+	}
+
+	// important to initialize to non-nil, empty value due to how fileIsIncluded works
+	skipDirs := skipList{}
+
+	for i, f := range zr.File {
+		f := f // make a copy for the Open closure
+		if err := ctx.Err(); err != nil {
+			return err // honor context cancellation
 		}
 
-		// make sure we do not copy the output file into the output
-		// file; that results in an infinite loop and disk exhaustion!
-		fpathAbs, err := filepath.Abs(fpath)
-		if err != nil {
-			return handleErr(fmt.Errorf("%s: getting absolute path: %v", fpath, err))
+		// ensure filename and comment are UTF-8 encoded (issue #147 and PR #305)
+		z.decodeText(&f.FileHeader)
+
+		if fileIsIncluded(skipDirs, f.Name) {
+			continue
 		}
-		if within(fpathAbs, destAbs) {
+
+		info := f.FileInfo()
+		file := FileInfo{
+			FileInfo:      info,
+			Header:        f.FileHeader,
+			NameInArchive: f.Name,
+			Open: func() (fs.File, error) {
+				openedFile, err := f.Open()
+				if err != nil {
+					return nil, err
+				}
+				return fileInArchive{openedFile, info}, nil
+			},
+		}
+
+		err := handleFile(ctx, file)
+		if errors.Is(err, fs.SkipAll) {
+			break
+		} else if errors.Is(err, fs.SkipDir) {
+			// if a directory, skip this path; if a file, skip the folder path
+			dirPath := f.Name
+			if !file.IsDir() {
+				dirPath = path.Dir(f.Name) + "/"
+			}
+			skipDirs.add(dirPath)
+		} else if err != nil {
+			if z.ContinueOnError {
+				log.Printf("[ERROR] %s: %v", f.Name, err)
+				continue
+			}
+			return fmt.Errorf("handling file %d: %s: %w", i, f.Name, err)
+		}
+	}
+
+	return nil
+}
+
+// decodeText decodes the name and comment fields from hdr into UTF-8.
+// It is a no-op if the text is already UTF-8 encoded or if z.TextEncoding
+// is not specified.
+func (z Zip) decodeText(hdr *zip.FileHeader) {
+	if hdr.NonUTF8 && z.TextEncoding != "" {
+		filename, err := decodeText(hdr.Name, z.TextEncoding)
+		if err == nil {
+			hdr.Name = filename
+		}
+		if hdr.Comment != "" {
+			comment, err := decodeText(hdr.Comment, z.TextEncoding)
+			if err == nil {
+				hdr.Comment = comment
+			}
+		}
+	}
+}
+
+// Insert appends the listed files into the provided Zip archive stream.
+func (z Zip) Insert(ctx context.Context, into io.ReadWriteSeeker, files []FileInfo) error {
+	// following very simple example at https://github.com/STARRY-S/zip?tab=readme-ov-file#usage
+	zu, err := szip.NewUpdater(into)
+	if err != nil {
+		return err
+	}
+	defer zu.Close()
+
+	for idx, file := range files {
+		if err := ctx.Err(); err != nil {
+			return err // honor context cancellation
+		}
+
+		hdr, err := szip.FileInfoHeader(file)
+		if err != nil {
+			return fmt.Errorf("getting info for file %d: %s: %w", idx, file.NameInArchive, err)
+		}
+		hdr.Name = file.NameInArchive // complete path, since FileInfoHeader() only has base name
+		if hdr.Name == "" {
+			hdr.Name = file.Name() // assume base name of file I guess
+		}
+
+		// customize header based on file properties
+		if file.IsDir() {
+			if !strings.HasSuffix(hdr.Name, "/") {
+				hdr.Name += "/" // required
+			}
+			hdr.Method = zip.Store
+		} else if z.SelectiveCompression {
+			// only enable compression on compressable files
+			ext := strings.ToLower(path.Ext(hdr.Name))
+			if _, ok := compressedFormats[ext]; ok {
+				hdr.Method = zip.Store
+			} else {
+				hdr.Method = z.Compression
+			}
+		}
+
+		w, err := zu.Append(hdr.Name, szip.APPEND_MODE_OVERWRITE)
+		if err != nil {
+			return fmt.Errorf("inserting file header: %d: %s: %w", idx, file.Name(), err)
+		}
+
+		// directories have no file body
+		if file.IsDir() {
 			return nil
 		}
-
-		// build the name to be used within the archive
-		nameInArchive, err := makeNameInArchive(sourceInfo, source, topLevelFolder, fpath)
-		if err != nil {
-			return handleErr(err)
-		}
-
-		var file io.ReadCloser
-		if info.Mode().IsRegular() {
-			file, err = os.Open(fpath)
-			if err != nil {
-				return handleErr(fmt.Errorf("%s: opening: %v", fpath, err))
-			}
-			defer file.Close()
-		}
-		err = z.Write(File{
-			FileInfo: FileInfo{
-				FileInfo:   info,
-				CustomName: nameInArchive,
-				SourcePath: fpath,
-			},
-			ReadCloser: file,
-		})
-		if err != nil {
-			return handleErr(fmt.Errorf("%s: writing: %s", fpath, err))
-		}
-
-		return nil
-	})
-}
-
-// Create opens z for writing a ZIP archive to out.
-func (z *Zip) Create(out io.Writer) error {
-	if z.zw != nil {
-		return fmt.Errorf("zip archive is already created for writing")
-	}
-	z.zw = zip.NewWriter(out)
-	if z.CompressionLevel != flate.DefaultCompression {
-		z.zw.RegisterCompressor(zip.Deflate, func(out io.Writer) (io.WriteCloser, error) {
-			return flate.NewWriter(out, z.CompressionLevel)
-		})
-	}
-	switch z.FileMethod {
-	case BZIP2:
-		z.zw.RegisterCompressor(uint16(BZIP2), func(out io.Writer) (io.WriteCloser, error) {
-			return bzip2.NewWriter(out, &bzip2.WriterConfig{Level: z.CompressionLevel})
-		})
-	case ZSTD:
-		z.zw.RegisterCompressor(uint16(ZSTD), func(out io.Writer) (io.WriteCloser, error) {
-			return zstd.NewWriter(out)
-		})
-	case XZ:
-		z.zw.RegisterCompressor(uint16(XZ), func(out io.Writer) (io.WriteCloser, error) {
-			return xz.NewWriter(out)
-		})
-	}
-	return nil
-}
-
-// Write writes f to z, which must have been opened for writing first.
-func (z *Zip) Write(f File) error {
-	if z.zw == nil {
-		return fmt.Errorf("zip archive was not created for writing first")
-	}
-	if f.FileInfo == nil {
-		return fmt.Errorf("no file info")
-	}
-	if f.FileInfo.Name() == "" {
-		return fmt.Errorf("missing file name")
-	}
-
-	header, err := zip.FileInfoHeader(f)
-	if err != nil {
-		return fmt.Errorf("%s: getting header: %v", f.Name(), err)
-	}
-
-	if f.IsDir() {
-		header.Name += "/" // required - strangely no mention of this in zip spec? but is in godoc...
-		header.Method = zip.Store
-	} else {
-		ext := strings.ToLower(path.Ext(header.Name))
-		if _, ok := compressedFormats[ext]; ok && z.SelectiveCompression {
-			header.Method = zip.Store
-		} else {
-			header.Method = uint16(z.FileMethod)
-		}
-	}
-
-	writer, err := z.zw.CreateHeader(header)
-	if err != nil {
-		return fmt.Errorf("%s: making header: %w", f.Name(), err)
-	}
-
-	return z.writeFile(f, writer)
-}
-
-func (z *Zip) writeFile(f File, writer io.Writer) error {
-	if f.IsDir() {
-		return nil // directories have no contents
-	}
-	if isSymlink(f) {
-		fi, ok := f.FileInfo.(FileInfo)
-		if !ok {
-			return fmt.Errorf("failed to cast fs.FileInfo to archiver.FileInfo: %v", f)
-		}
-		// file body for symlinks is the symlink target
-		linkTarget, err := os.Readlink(fi.SourcePath)
-		if err != nil {
-			return fmt.Errorf("%s: readlink: %v", fi.SourcePath, err)
-		}
-		_, err = writer.Write([]byte(filepath.ToSlash(linkTarget)))
-		if err != nil {
-			return fmt.Errorf("%s: writing symlink target: %v", fi.SourcePath, err)
-		}
-		return nil
-	}
-
-	if f.ReadCloser == nil {
-		return fmt.Errorf("%s: no way to read file contents", f.Name())
-	}
-	_, err := io.Copy(writer, f)
-	if err != nil {
-		return fmt.Errorf("%s: copying contents: %w", f.Name(), err)
-	}
-
-	return nil
-}
-
-// Open opens z for reading an archive from in,
-// which is expected to have the given size and
-// which must be an io.ReaderAt.
-func (z *Zip) Open(in io.Reader, size int64) error {
-	inRdrAt, ok := in.(io.ReaderAt)
-	if !ok {
-		return fmt.Errorf("reader must be io.ReaderAt")
-	}
-	if z.zr != nil {
-		return fmt.Errorf("zip archive is already open for reading")
-	}
-	var err error
-	z.zr, err = zip.NewReader(inRdrAt, size)
-	if err != nil {
-		return fmt.Errorf("creating reader: %v", err)
-	}
-	registerDecompressor(z.zr)
-	z.ridx = 0
-	return nil
-}
-
-// Read reads the next file from z, which must have
-// already been opened for reading. If there are no
-// more files, the error is io.EOF. The File must
-// be closed when finished reading from it.
-func (z *Zip) Read() (File, error) {
-	if z.zr == nil {
-		return File{}, fmt.Errorf("zip archive is not open")
-	}
-	if z.ridx >= len(z.zr.File) {
-		return File{}, io.EOF
-	}
-
-	// access the file and increment counter so that
-	// if there is an error processing this file, the
-	// caller can still iterate to the next file
-	zf := z.zr.File[z.ridx]
-	z.ridx++
-
-	file := File{
-		FileInfo: zf.FileInfo(),
-		Header:   zf.FileHeader,
-	}
-
-	rc, err := zf.Open()
-	if err != nil {
-		return file, fmt.Errorf("%s: open compressed file: %v", zf.Name, err)
-	}
-	file.ReadCloser = rc
-
-	return file, nil
-}
-
-// Close closes the zip archive(s) opened by Create and Open.
-func (z *Zip) Close() error {
-	if z.zr != nil {
-		z.zr = nil
-	}
-	if z.zw != nil {
-		zw := z.zw
-		z.zw = nil
-		return zw.Close()
-	}
-	return nil
-}
-
-// Walk calls walkFn for each visited item in archive.
-func (z *Zip) Walk(archive string, walkFn WalkFunc) error {
-	zr, err := zip.OpenReader(archive)
-	if err != nil {
-		return fmt.Errorf("opening zip reader: %v", err)
-	}
-	defer zr.Close()
-	registerDecompressor(&zr.Reader)
-	for _, zf := range zr.File {
-		zfrc, err := zf.Open()
-		if err != nil {
-			if zfrc != nil {
-				zfrc.Close()
-			}
-			if z.ContinueOnError {
-				log.Printf("[ERROR] Opening %s: %v", zf.Name, err)
+		if err := openAndCopyFile(file, w); err != nil {
+			if z.ContinueOnError && ctx.Err() == nil {
+				log.Printf("[ERROR] appending file %d into archive: %s: %v", idx, file.Name(), err)
 				continue
 			}
-			return fmt.Errorf("opening %s: %v", zf.Name, err)
-		}
-
-		err = walkFn(File{
-			FileInfo:   zf.FileInfo(),
-			Header:     zf.FileHeader,
-			ReadCloser: zfrc,
-		})
-		zfrc.Close()
-		if err != nil {
-			if err == ErrStopWalk {
-				break
-			}
-			if z.ContinueOnError {
-				log.Printf("[ERROR] Walking %s: %v", zf.Name, err)
-				continue
-			}
-			return fmt.Errorf("walking %s: %v", zf.Name, err)
+			return fmt.Errorf("copying inserted file %d: %s: %w", idx, file.Name(), err)
 		}
 	}
 
 	return nil
 }
 
-// Extract extracts a single file from the zip archive.
-// If the target is a directory, the entire folder will
-// be extracted into destination.
-func (z *Zip) Extract(source, target, destination string) error {
-	// target refers to a path inside the archive, which should be clean also
-	target = path.Clean(target)
-
-	// if the target ends up being a directory, then
-	// we will continue walking and extracting files
-	// until we are no longer within that directory
-	var targetDirPath string
-
-	return z.Walk(source, func(f File) error {
-		zfh, ok := f.Header.(zip.FileHeader)
-		if !ok {
-			return fmt.Errorf("expected header to be zip.FileHeader but was %T", f.Header)
-		}
-
-		// importantly, cleaning the path strips tailing slash,
-		// which must be appended to folders within the archive
-		name := path.Clean(zfh.Name)
-		if f.IsDir() && target == name {
-			targetDirPath = path.Dir(name)
-		}
-
-		if within(target, zfh.Name) {
-			// either this is the exact file we want, or is
-			// in the directory we want to extract
-
-			// build the filename we will extract to
-			end, err := filepath.Rel(targetDirPath, zfh.Name)
-			if err != nil {
-				return fmt.Errorf("relativizing paths: %v", err)
-			}
-			joined := filepath.Join(destination, end)
-
-			err = z.extractFile(f, joined, &zfh)
-			if err != nil {
-				return fmt.Errorf("extracting file %s: %v", zfh.Name, err)
-			}
-
-			// if our target was not a directory, stop walk
-			if targetDirPath == "" {
-				return ErrStopWalk
-			}
-		} else if targetDirPath != "" {
-			// finished walking the entire directory
-			return ErrStopWalk
-		}
-
-		return nil
-	})
+type seekReaderAt interface {
+	io.ReaderAt
+	io.Seeker
 }
 
-// Match returns true if the format of file matches this
-// type's format. It should not affect reader position.
-func (*Zip) Match(file io.ReadSeeker) (bool, error) {
-	currentPos, err := file.Seek(0, io.SeekCurrent)
-	if err != nil {
-		return false, err
-	}
-	_, err = file.Seek(0, 0)
-	if err != nil {
-		return false, err
-	}
-	defer func() {
-		_, _ = file.Seek(currentPos, io.SeekStart)
-	}()
-
-	buf := make([]byte, 4)
-	if n, err := file.Read(buf); err != nil || n < 4 {
-		return false, nil
-	}
-	return bytes.Equal(buf, []byte("PK\x03\x04")), nil
-}
-
-func (z *Zip) String() string { return "zip" }
-
-// NewZip returns a new, default instance ready to be customized and used.
-func NewZip() *Zip {
-	return &Zip{
-		CompressionLevel:     flate.DefaultCompression,
-		MkdirAll:             true,
-		SelectiveCompression: true,
-		FileMethod:           Deflate,
-	}
-}
-
-// Compile-time checks to ensure type implements desired interfaces.
-var (
-	_ = Reader(new(Zip))
-	_ = Writer(new(Zip))
-	_ = Archiver(new(Zip))
-	_ = Unarchiver(new(Zip))
-	_ = Walker(new(Zip))
-	_ = Extractor(new(Zip))
-	_ = Matcher(new(Zip))
-	_ = ExtensionChecker(new(Zip))
-	_ = FilenameChecker(new(Zip))
+// Additional compression methods not offered by archive/zip.
+// See https://pkware.cachefly.net/webdocs/casestudies/APPNOTE.TXT section 4.4.5.
+const (
+	ZipMethodBzip2 = 12
+	// TODO: LZMA: Disabled - because 7z isn't able to unpack ZIP+LZMA ZIP+LZMA2 archives made this way - and vice versa.
+	// ZipMethodLzma     = 14
+	ZipMethodZstd = 93
+	ZipMethodXz   = 95
 )
 
 // compressedFormats is a (non-exhaustive) set of lowercased
 // file extensions for formats that are typically already
 // compressed. Compressing files that are already compressed
-// is inefficient, so use this set of extension to avoid that.
+// is inefficient, so use this set of extensions to avoid that.
 var compressedFormats = map[string]struct{}{
 	".7z":   {},
 	".avi":  {},
@@ -707,5 +383,61 @@ var compressedFormats = map[string]struct{}{
 	".zipx": {},
 }
 
-// DefaultZip is a default instance that is conveniently ready to use.
-var DefaultZip = NewZip()
+var encodings = map[string]encoding.Encoding{
+	"ibm866":            charmap.CodePage866,
+	"iso8859_2":         charmap.ISO8859_2,
+	"iso8859_3":         charmap.ISO8859_3,
+	"iso8859_4":         charmap.ISO8859_4,
+	"iso8859_5":         charmap.ISO8859_5,
+	"iso8859_6":         charmap.ISO8859_6,
+	"iso8859_7":         charmap.ISO8859_7,
+	"iso8859_8":         charmap.ISO8859_8,
+	"iso8859_8I":        charmap.ISO8859_8I,
+	"iso8859_10":        charmap.ISO8859_10,
+	"iso8859_13":        charmap.ISO8859_13,
+	"iso8859_14":        charmap.ISO8859_14,
+	"iso8859_15":        charmap.ISO8859_15,
+	"iso8859_16":        charmap.ISO8859_16,
+	"koi8r":             charmap.KOI8R,
+	"koi8u":             charmap.KOI8U,
+	"macintosh":         charmap.Macintosh,
+	"windows874":        charmap.Windows874,
+	"windows1250":       charmap.Windows1250,
+	"windows1251":       charmap.Windows1251,
+	"windows1252":       charmap.Windows1252,
+	"windows1253":       charmap.Windows1253,
+	"windows1254":       charmap.Windows1254,
+	"windows1255":       charmap.Windows1255,
+	"windows1256":       charmap.Windows1256,
+	"windows1257":       charmap.Windows1257,
+	"windows1258":       charmap.Windows1258,
+	"macintoshcyrillic": charmap.MacintoshCyrillic,
+	"gbk":               simplifiedchinese.GBK,
+	"gb18030":           simplifiedchinese.GB18030,
+	"big5":              traditionalchinese.Big5,
+	"eucjp":             japanese.EUCJP,
+	"iso2022jp":         japanese.ISO2022JP,
+	"shiftjis":          japanese.ShiftJIS,
+	"euckr":             korean.EUCKR,
+	"utf16be":           unicode.UTF16(unicode.BigEndian, unicode.IgnoreBOM),
+	"utf16le":           unicode.UTF16(unicode.LittleEndian, unicode.IgnoreBOM),
+}
+
+// decodeText returns UTF-8 encoded text from the given charset.
+// Thanks to @zxdvd for contributing non-UTF-8 encoding logic in
+// #149, and to @pashifika for helping in #305.
+func decodeText(input, charset string) (string, error) {
+	if enc, ok := encodings[charset]; ok {
+		return enc.NewDecoder().String(input)
+	}
+	return "", fmt.Errorf("unrecognized charset %s", charset)
+}
+
+var zipHeader = []byte("PK\x03\x04") // NOTE: headers of empty zip files might end with 0x05,0x06 or 0x06,0x06 instead of 0x03,0x04
+
+// Interface guards
+var (
+	_ Archiver      = Zip{}
+	_ ArchiverAsync = Zip{}
+	_ Extractor     = Zip{}
+)
