@@ -1,76 +1,89 @@
 package archiver
 
 import (
-	"fmt"
+	"bytes"
+	"context"
 	"io"
-	"path/filepath"
+	"strings"
 
 	"github.com/klauspost/compress/gzip"
 	"github.com/klauspost/pgzip"
 )
 
+func init() {
+	RegisterFormat(Gz{})
+}
+
 // Gz facilitates gzip compression.
 type Gz struct {
+	// Gzip compression level. See https://pkg.go.dev/compress/flate#pkg-constants
+	// for some predefined constants. If 0, DefaultCompression is assumed rather
+	// than no compression.
 	CompressionLevel int
-	SingleThreaded   bool
+
+	// DisableMultistream controls whether the reader supports multistream files.
+	// See https://pkg.go.dev/compress/gzip#example-Reader.Multistream
+	DisableMultistream bool
+
+	// Use a fast parallel Gzip implementation. This is only
+	// effective for large streams (about 1 MB or greater).
+	Multithreaded bool
 }
 
-// Compress reads in, compresses it, and writes it to out.
-func (gz *Gz) Compress(in io.Reader, out io.Writer) error {
-	var w io.WriteCloser
-	var err error
-	if gz.SingleThreaded {
-		w, err = gzip.NewWriterLevel(out, gz.CompressionLevel)
-	} else {
-		w, err = pgzip.NewWriterLevel(out, gz.CompressionLevel)
+func (Gz) Extension() string { return ".gz" }
+
+func (gz Gz) Match(_ context.Context, filename string, stream io.Reader) (MatchResult, error) {
+	var mr MatchResult
+
+	// match filename
+	if strings.Contains(strings.ToLower(filename), gz.Extension()) {
+		mr.ByName = true
 	}
+
+	// match file header
+	buf, err := readAtMost(stream, len(gzHeader))
 	if err != nil {
-		return err
+		return mr, err
 	}
-	defer w.Close()
-	_, err = io.Copy(w, in)
-	return err
+	mr.ByStream = bytes.Equal(buf, gzHeader)
+
+	return mr, nil
 }
 
-// Decompress reads in, decompresses it, and writes it to out.
-func (gz *Gz) Decompress(in io.Reader, out io.Writer) error {
-	var r io.ReadCloser
+func (gz Gz) OpenWriter(w io.Writer) (io.WriteCloser, error) {
+	// assume default compression level if 0, rather than no
+	// compression, since no compression on a gzipped file
+	// doesn't make any sense in our use cases
+	level := gz.CompressionLevel
+	if level == 0 {
+		level = gzip.DefaultCompression
+	}
+
+	var wc io.WriteCloser
 	var err error
-	if gz.SingleThreaded {
-		r, err = gzip.NewReader(in)
+	if gz.Multithreaded {
+		wc, err = pgzip.NewWriterLevel(w, level)
 	} else {
-		r, err = pgzip.NewReader(in)
+		wc, err = gzip.NewWriterLevel(w, level)
 	}
-	if err != nil {
-		return err
-	}
-	defer r.Close()
-	_, err = io.Copy(out, r)
-	return err
+	return wc, err
 }
 
-// CheckExt ensures the file extension matches the format.
-func (gz *Gz) CheckExt(filename string) error {
-	if filepath.Ext(filename) != ".gz" {
-		return fmt.Errorf("filename must have a .gz extension")
+func (gz Gz) OpenReader(r io.Reader) (io.ReadCloser, error) {
+	if gz.Multithreaded {
+		gzR, err := pgzip.NewReader(r)
+		if gzR != nil && gz.DisableMultistream {
+			gzR.Multistream(false)
+		}
+		return gzR, err
 	}
-	return nil
+
+	gzR, err := gzip.NewReader(r)
+	if gzR != nil && gz.DisableMultistream {
+		gzR.Multistream(false)
+	}
+	return gzR, err
 }
 
-func (gz *Gz) String() string { return "gz" }
-
-// NewGz returns a new, default instance ready to be customized and used.
-func NewGz() *Gz {
-	return &Gz{
-		CompressionLevel: gzip.DefaultCompression,
-	}
-}
-
-// Compile-time checks to ensure type implements desired interfaces.
-var (
-	_ = Compressor(new(Gz))
-	_ = Decompressor(new(Gz))
-)
-
-// DefaultGz is a default instance that is conveniently ready to use.
-var DefaultGz = NewGz()
+// magic number at the beginning of gzip files
+var gzHeader = []byte{0x1f, 0x8b}
