@@ -1,58 +1,54 @@
 package archiver
 
 import (
-	"fmt"
+	"bytes"
+	"context"
 	"io"
-	"path/filepath"
+	"strings"
 
+	fastxz "github.com/therootcompany/xz"
 	"github.com/ulikunitz/xz"
-	fastxz "github.com/xi2/xz"
 )
 
-// Xz facilitates XZ compression.
+func init() {
+	RegisterFormat(Xz{})
+}
+
+// Xz facilitates xz compression.
 type Xz struct{}
 
-// Compress reads in, compresses it, and writes it to out.
-func (x *Xz) Compress(in io.Reader, out io.Writer) error {
-	w, err := xz.NewWriter(out)
+func (Xz) Extension() string { return ".xz" }
+
+func (x Xz) Match(_ context.Context, filename string, stream io.Reader) (MatchResult, error) {
+	var mr MatchResult
+
+	// match filename
+	if strings.Contains(strings.ToLower(filename), x.Extension()) {
+		mr.ByName = true
+	}
+
+	// match file header
+	buf, err := readAtMost(stream, len(xzHeader))
 	if err != nil {
-		return err
+		return mr, err
 	}
-	defer w.Close()
-	_, err = io.Copy(w, in)
-	return err
+	mr.ByStream = bytes.Equal(buf, xzHeader)
+
+	return mr, nil
 }
 
-// Decompress reads in, decompresses it, and writes it to out.
-func (x *Xz) Decompress(in io.Reader, out io.Writer) error {
-	r, err := fastxz.NewReader(in, 0)
+func (Xz) OpenWriter(w io.Writer) (io.WriteCloser, error) {
+	return xz.NewWriter(w)
+}
+
+func (Xz) OpenReader(r io.Reader) (io.ReadCloser, error) {
+	xr, err := fastxz.NewReader(r, 0)
 	if err != nil {
-		return err
+		return nil, err
 	}
-	_, err = io.Copy(out, r)
-	return err
+	return io.NopCloser(xr), err
 }
 
-// CheckExt ensures the file extension matches the format.
-func (x *Xz) CheckExt(filename string) error {
-	if filepath.Ext(filename) != ".xz" {
-		return fmt.Errorf("filename must have a .xz extension")
-	}
-	return nil
-}
-
-func (x *Xz) String() string { return "xz" }
-
-// NewXz returns a new, default instance ready to be customized and used.
-func NewXz() *Xz {
-	return new(Xz)
-}
-
-// Compile-time checks to ensure type implements desired interfaces.
-var (
-	_ = Compressor(new(Xz))
-	_ = Decompressor(new(Xz))
-)
-
-// DefaultXz is a default instance that is conveniently ready to use.
-var DefaultXz = NewXz()
+// magic number at the beginning of xz files; see section 2.1.1.1
+// of https://tukaani.org/xz/xz-file-format.txt
+var xzHeader = []byte{0xfd, 0x37, 0x7a, 0x58, 0x5a, 0x00}
